pyproject.toml
[project]
name = "experimentation-benchmark"
version = "0.1.0"
description = "Standalone benchmark comparing on-demand vs pre-aggregation SQL approaches for A/B test analysis"
requires-python = ">=3.9"
dependencies = [
    "duckdb>=1.0",
    "jinja2>=3.1",
    "pandas>=2.0",
    "pyyaml>=6.0",
]

[project.optional-dependencies]
postgres = ["psycopg2-binary>=2.9"]
dev = ["black", "ruff"]

[tool.setuptools.packages.find]
include = ["benchmark*", "data*"]
